Add files via upload
notes:
C Lang has a thing called `Struct`. I wanted to use Structs for xy positions.

Initially I was keeping the guard position (variable `gp`) as a tuple of two integers with the first and second numbers as x and y.

to get those values separately it would be `gp[0]` and `gp[1]`.
but as an object it would be `gp.x` and `gp.y` which I find more readable than as a tuple 👌 

Python does not directly have structs but I can approximate that by creating a class with only members, no methods.
The two members being of course x and y.